session: update to only clear tokens given, if any

+++ storage.go
@@ -89,6 +89,29 @@ func (j *Jeff) store(ctx context.Context, s Session) error {
 	return j.s.Store(ctx, s.Key, bts, now().Add(24*30*time.Hour))
 }
 
-func (j *Jeff) clear(ctx context.Context, key []byte) error {
-	return j.s.Delete(ctx, key)
+func (j *Jeff) clear(ctx context.Context, key []byte, tokens ...[]byte) error {
+	if len(tokens) == 0 {
+		return j.s.Delete(ctx, key)
+	}
+
+	sl, err := j.load(ctx, key)
+	if err != nil {
+		return err
+	}
+
+	// if it's found, remove it
+	for _, tok := range tokens {
+		if _, i := find(sl, tok); i >= 0 {
+			sl = append(sl[:i], sl[i+1:]...)
+		}
+	}
+
+	// prune expired sessions
+	sl = prune(sl)
+	bts, err := sl.MarshalMsg(nil)
+	if err != nil {
+		return err
+	}
+	// Global Expiration 30d, TODO: make configurable
+	return j.s.Store(ctx, key, bts, now().Add(24*30*time.Hour))
 }
